+++ dev/integration_tests/android_views/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_driver/driver_extension.dart';
-import 'motion_event_page.dart';
+import 'motion_events_page.dart';
 import 'page.dart';
 
 final List<Page> _allPages = <Page>[
